+++ merlin/config/configfile.py
@@ -129,21 +129,20 @@ def get_config(path):
     return config
 
 
-def load_default_timeout(config):
-    seconds = 60 * 60 * 24
+def load_default_celery(config):
     try:
         config["celery"]
     except KeyError:
         config["celery"] = {}
     try:
-        config["celery"]["visibility_timeout_seconds"]
+        config["celery"]["override"]
     except KeyError:
-        config["celery"]["visibility_timeout_seconds"] = seconds
+        config["celery"]["override"] = None
 
 
 def load_defaults(config):
     load_default_user_names(config)
-    load_default_timeout(config)
+    load_default_celery(config)
 
 
 def is_debug():